First Creation of index
I used to bootstrap css and an additional one to override it. Used a template from w3 schools to create the layout.
I added in my own images and used circles below for three seperate coloums, again with own images. Changed the text on the navigation bar. Created a social media bar at the bottom.
There is still problems with the sizing of the images on the scrolling bar. Need to fix the box of the rows so that the circle images into it. Need to add links to other pages. Not happy with the size of the instagram button.